VFB-412: refactorized the code, made the function common between clients and parcels
